test(e2e-tests): refactored e2e tests for consistency and flexibility
