+++ e2e-tests/README.md
@@ -24,11 +24,17 @@ or like this to select a specific feature file:
 node_modules/cypress/bin/cypress run --headed -b chrome --env demoDelay=1000 --spec cypress/integration/appeal-statement-submission.feature
 ```
 
+Evidence for FQC can be generated by running the tests or a subset of them against any accessible environment e.g.
+```
+node_modules/cypress/bin/cypress run --headed -b chrome --config baseUrl="https://appeals-dev.planninginspectorate.gov.uk/" --spec "cypress/integration/appellant-submission-your-details-*.feature" --env "TAGS='not @wip',demoDelay=1000"
+```
+
+
 #### Test data
 
 The maximum permitted file size for uploading is configurable with a default value of 50MB. 
-A script `create-large-test-files.sh` is available to create files to support testing. This will look for an env var 
-`FILE_UPLOAD_MAX_FILE_SIZE_BYTES` and will use the value there if present instead of the default.
+A script `create-large-test-files.sh` is available to create files to support testing and should be run automatically 
+using `make install`. This will look for an env var `FILE_UPLOAD_MAX_FILE_SIZE_BYTES` and will use the value there if present instead of the default.
 The files named `appeal-statement-invalid-too-big.png` and `appeal-statement-valid-max-size.png` will be placed in 
 the `cypress/fixtures` folder. This script runs as part of the CI sequence to ensure that the files are available 
 for e2e tests in the pipelines. To create the required files locally run this command once from the `e2e-tests` folder:
