--- e2e-tests/cypress/support/appellant-submission-applicant-name/goToApplicantNamePage.js
@@ -1,5 +1,0 @@
-module.exports = () => {
-  // go to the right page
-  cy.visit('/appellant-submission/applicant-name');
-  cy.wait(Cypress.env('demoDelay'));
-};
