--- e2e-tests/cypress/support/appellant-submission-who-are-you/confirmOriginalAppellantNotAsked.js
@@ -1,7 +1,0 @@
-module.exports = () => {
-  // confirm we are in the right place
-  cy.url().should('include', '/appellant-submission/task-list');
-
-  // pause long enough to capture a nice video
-  cy.wait(Cypress.env('demoDelay'));
-};
