--- e2e-tests/cypress/support/appellant-submission-who-are-you/goToWhoAreYouPage.js
@@ -1,5 +1,0 @@
-module.exports = () => {
-  // go to the right page
-  cy.visit('/appellant-submission/who-are-you');
-  cy.wait(Cypress.env('demoDelay'));
-};
