+++ e2e-tests/cypress/support/appellant-submission-your-details/clickSaveAndContinue.js
