+++ e2e-tests/cypress/support/appellant-submission-your-details/confirmAnswered.js
@@ -1,9 +1,7 @@
 module.exports = (answer) => {
   cy.goToWhoAreYouPage();
-
   cy.get('[data-cy="answer-' + answer + '"]')
     .first()
     .should('be.checked');
-
   cy.wait(Cypress.env('demoDelay'));
 };
