+++ e2e-tests/cypress/support/appellant-submission-your-details/confirmNavigationApplicantNamePage.js
@@ -1,7 +1,4 @@
 module.exports = () => {
-  // confirm we are in the right place
   cy.url().should('include', '/appellant-submission/applicant-name');
-
-  // pause long enough to capture a nice video
   cy.wait(Cypress.env('demoDelay'));
-};
+}
