+++ e2e-tests/cypress/support/appellant-submission-your-details/confirmNavigationTaskListPage.js
@@ -1,4 +1,4 @@
-module.exports = (name, email) => {
+module.exports = () => {
   cy.url().should('include', '/appellant-submission/applicant-name');
   cy.wait(Cypress.env('demoDelay'));
-};
+}
