+++ e2e-tests/cypress/support/appellant-submission-your-details/confirmNavigationWhoAreYouPage.js
@@ -1,0 +1,4 @@
+module.exports = () => {
+  cy.url().should('include', '/appellant-submission/who-are-you');
+  cy.wait(Cypress.env('demoDelay'));
+}
