+++ e2e-tests/cypress/support/appellant-submission-your-details/confirmNavigationYourDetailsPage.js
@@ -1,0 +1,4 @@
+module.exports = () => {
+  cy.url().should('include', '/appellant-submission/your-details');
+  cy.wait(Cypress.env('demoDelay'));
+}
