+++ e2e-tests/cypress/support/appellant-submission-your-details/confirmOriginalAppellantNotAsked.js
@@ -1,5 +1,4 @@
 module.exports = () => {
-  cy.get('#are-you-the-original-appellant').click();
-
+  cy.url().should('include', '/appellant-submission/task-list');
   cy.wait(Cypress.env('demoDelay'));
 };
