+++ e2e-tests/cypress/support/appellant-submission-your-details/confirmOriginalApplicantWasRejected.js
@@ -1,7 +1,5 @@
 module.exports = (errorMessage) => {
-  // confirm we are in the right place
   cy.url().should('include', '/appellant-submission/applicant-name');
-
   cy.get('.govuk-error-summary__list')
     .invoke('text')
     .then((text) => {
@@ -10,6 +8,5 @@ module.exports = (errorMessage) => {
       }
       errorMessage.forEach((errorMessage) => expect(text).to.contain(errorMessage));
     });
-  // pause long enough to capture a nice video
   cy.wait(Cypress.env('demoDelay'));
 };
