+++ e2e-tests/cypress/support/appellant-submission-your-details/provideAnswerYes.js
@@ -1,0 +1,4 @@
+module.exports = () => {
+  cy.get('#are-you-the-original-appellant').click();
+  cy.wait(Cypress.env('demoDelay'));
+};
