+++ e2e-tests/cypress/support/appellant-submission-your-details/provideApplicantName.js
@@ -1,6 +1,4 @@
 module.exports = (name) => {
-  // provide the name
   cy.get('#behalf-appellant-name').type(`{selectall}{backspace}${name}`);
-
   cy.wait(Cypress.env('demoDelay'));
 };
